+++ src/components/@shared/Price/Conversion.tsx
@@ -10,13 +10,11 @@ const cx = classNames.bind(styles)
 export default function Conversion({
   price,
   className,
-  hideApproximateSymbol,
-  showTVLLabel
+  hideApproximateSymbol
 }: {
   price: string // expects price in OCEAN, not wei
   className?: string
   hideApproximateSymbol?: boolean
-  showTVLLabel?: boolean
 }): ReactElement {
   const { prices } = usePrices()
   const { currency, locale } = useUserPreferences()
@@ -29,7 +27,6 @@ export default function Conversion({
 
   const styleClasses = cx({
     conversion: true,
-    removeTvlPadding: showTVLLabel,
     [className]: className
   })
 
@@ -39,7 +36,7 @@ export default function Conversion({
       return
     }
 
-    const conversionValue = (prices as any)[currency.toLowerCase()]
+    const conversionValue = prices[currency.toLowerCase()]
     const converted = conversionValue * Number(price)
     const convertedFormatted = formatCurrency(
       converted,
@@ -64,7 +61,6 @@ export default function Conversion({
       className={styleClasses}
       title="Approximation based on current OCEAN spot price on Coingecko"
     >
-      {showTVLLabel && 'TVL'}
       {!hideApproximateSymbol && '≈ '}
       <strong dangerouslySetInnerHTML={{ __html: priceConverted }} />{' '}
       {!isFiat && currency}
