+++ src/components/@shared/atoms/Tooltip.module.css
@@ -14,8 +14,8 @@
 }
 
 .icon {
-  width: 1em;
-  height: 1em;
+  width: 0.85em;
+  height: 0.85em;
   cursor: help;
   display: inline-block;
   margin-bottom: -0.1em;
