+++ src/components/Asset/AssetActions/Pool/Add/Output.module.css
@@ -1,7 +1,4 @@
 .output {
-  display: grid;
-  gap: var(--spacer);
-  grid-template-columns: 1fr 1fr;
   padding-bottom: calc(var(--spacer) / 2);
 }
 
@@ -9,14 +6,6 @@
   font-weight: var(--font-weight-bold);
   margin-bottom: calc(var(--spacer) / 8);
   font-size: var(--font-size-small);
-}
-
-.output div:first-child [class*='token'] {
-  white-space: normal;
-}
-
-.output div:first-child [class*='token'] > figure {
-  display: none;
 }
 
 .help {
