+++ src/components/Asset/AssetActions/Pool/TokenList.module.css
@@ -20,7 +20,7 @@
 
 .tokens {
   display: grid;
-  grid-template-columns: 1fr 1fr;
+  grid-template-columns: repeat(auto-fill, minmax(10rem, 1fr));
 }
 
 .title {
