+++ src/components/Asset/AssetActions/Pool/TokenList.tsx
@@ -11,42 +11,33 @@ export default function TokenList({
   baseTokenSymbol,
   datatokenValue,
   datatokenSymbol,
-  poolShares,
   conversion,
-  highlight,
-  showTVLLabel
+  highlight
 }: {
   title: string | ReactNode
-  children: ReactNode
+  children?: ReactNode
   baseTokenValue: string
   baseTokenSymbol: string
-  datatokenValue: string
-  datatokenSymbol: string
-  poolShares: string
+  datatokenValue?: string
+  datatokenSymbol?: string
   conversion: Decimal
   highlight?: boolean
-  showTVLLabel?: boolean
 }): ReactElement {
   return (
     <div className={`${styles.tokenlist} ${highlight ? styles.highlight : ''}`}>
       <h3 className={styles.title}>{title}</h3>
       <div className={styles.tokens}>
-        <div>
-          <Token symbol={baseTokenSymbol} balance={baseTokenValue} />
+        <Token symbol={baseTokenSymbol} balance={baseTokenValue} />
+        {datatokenValue && (
           <Token symbol={datatokenSymbol} balance={datatokenValue} />
-          {conversion.greaterThan(0) && (
-            <Conversion
-              price={conversion.toString()}
-              className={styles.totalLiquidity}
-              showTVLLabel={showTVLLabel}
-            />
-          )}
-        </div>
-
-        <div>
-          <Token symbol="pool shares" balance={poolShares} noIcon />
-          {children}
-        </div>
+        )}
+        {conversion.greaterThan(0) && (
+          <Conversion
+            price={conversion.toString()}
+            className={styles.totalLiquidity}
+          />
+        )}
+        {children}
       </div>
     </div>
   )
